requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.0.0
anthropic>=0.25.0
pytesseract>=0.3.10
opencv-python>=4.9.0
Pillow>=10.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
Werkzeug>=3.0.0
